feat: add a option to disable the lock mechanism

+++ experimentor/track_log.py
@@ -6,7 +6,9 @@ class TrackLog:
 
     On initialization, it will create a lock file to make sure that only one
     process is using the directory. The lock file will be removed when the
-    object is deleted.
+    object is deleted. You can disable the lock by setting the disable_lock
+    parameter to True when initializing the object. In this case, there
+    can be multiple processes using the directory at the same time.
 
     When running experiments, a log file will be created for each experiment.
     The file name is the current time in the format of '%Y_%m_%d_%H_%M_%S.log'
@@ -17,16 +19,26 @@ class TrackLog:
     because improperly changing the behavior might lead to some unexpected
     results.
     """
-    def __init__(self, root_dir: str):
+    def __init__(self, root_dir: str, disable_lock=False):
+        """
+        Initialize the TrackLog object.
+        :param root_dir: The root directory to store the log files.
+        :param disable_lock: Whether to disable the lock file used to
+         guarantee that only one process is using the directory.
+        """
         self.root_dir = root_dir
+        self.disable_lock = disable_lock
         self.init_dir()
 
     def __del__(self):
-        lock_file = os.path.join(self.root_dir, 'lock')
-        os.remove(lock_file)
+        if not self.disable_lock:
+            lock_file = os.path.join(self.root_dir, 'lock')
+            os.remove(lock_file)
 
     def init_dir(self):
         os.makedirs(self.root_dir, exist_ok=True)
+        if self.disable_lock:
+            return
 
         # Make sure there is no other process using the directory
         lock_file = os.path.join(self.root_dir, 'lock')
